requirements.txt
pandas
python-dotenv
openai
jinja2 
markdown2